moved mockData to an own folder
